optscale-deploy/requirements.txt
ansible==2.9.27
PyYAML==5.3.1
kubernetes==10.0.1
docker==3.0.1
